calendar final

+++ app/Http/Controllers/BookingController.php
@@ -3,9 +3,9 @@
 namespace App\Http\Controllers;
 
 
-use App\Booking;
 use App\User;
 use Illuminate\Http\Request;
+use App\Booking;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
 
@@ -31,7 +31,10 @@ class BookingController extends Controller
         $booking->save();
         return response()->json(['success'=>true, 'data' => $booking]);
     }
-    public function delete(Request $request) {
+    public function delete(Request $request){
+        if($request['data']['author']!== Auth::user()->ISN) {
+            return response()->json(['success'=>false, 'message'=>'Невозможно удалить!']);
+        }
         $booking = Booking::where('id', ($request['data']['id']))->delete();
         return response()->json(['success'=>true ]);
     }
